+++ include/toltrig/toltrig.hpp
@@ -1,4 +1,5 @@
 #pragma once
 #include "constants.hpp"
 #include "cos.hpp"
+#include "sin.hpp"
 #include "tan.hpp"